Moved other dialogs into the new settings dialog file and their according state
